# docs/devicerepository_swagger.yaml
basePath: /
definitions:
...
  models.Attribute:
    properties:
      key:
        type: string
      origin:
        type: string
      value:
        type: string
    type: object
...
  models.Device:
    properties:
      attributes:
        items:
          $ref: '#/definitions/models.Attribute'
        type: array
      device_type_id:
        type: string
      id:
        type: string
      local_id:
        type: string
      name:
        type: string
      owner_id:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Device-Repository API
  version: "0.1"
paths:
...
  /devices:
    delete:
      description: delete multiple devices
      parameters:
      - description: ids to be deleted
        in: body
        name: message
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: delete multiple devices
      tags:
      - devices
    get:
      description: list devices
      parameters:
      - description: default 100, will be ignored if 'ids' is set
        in: query
        name: limit
        type: integer
      - description: default 0, will be ignored if 'ids' is set
        in: query
        name: offset
        type: integer
      - description: filter
        in: query
        name: search
        type: string
      - description: default name.asc
        in: query
        name: sort
        type: string
      - description: filter; ignores limit/offset; comma-separated list
        in: query
        name: ids
        type: string
      - description: in combination with owner; fills ids filter; comma-separated
          list
        in: query
        name: local_ids
        type: string
      - description: used in combination with local_ids to fill ids filter; defaults
          to requesting user
        in: query
        name: owner
        type: string
      - description: filter; comma-separated list
        in: query
        name: device-type-ids
        type: string
      - description: filter; comma-separated list; lists elements only if they have
          an attribute key that is in the given list
        in: query
        name: attr-keys
        type: string
      - description: filter; comma-separated list; lists elements only if they have
          an attribute value that is in the given list
        in: query
        name: attr-values
        type: string
      - description: filter; valid values are 'online', 'offline' and an empty string
          for unknown states
        in: query
        name: connection-state
        type: integer
      - description: default 'r'; used to check permissions on request; valid values
          are 'r', 'w', 'x', 'a' for read, write, execute, administrate
        in: query
        name: p
        type: string
      - description: JSON encoded []models.Attribute, attribute value and origin will
          only be checked if set, otherwise all values or origins will be blacklisted
        in: query
        name: device-attribute-blacklist
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Device'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: list devices
      tags:
      - devices
    post:
      description: create device
      parameters:
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: create device
      tags:
      - devices
    put:
      consumes:
      - application/json
      description: validate device
      parameters:
      - description: must be true; reminder, that this is not an update but a validation
        in: query
        name: dry-run
        required: true
        type: boolean
      - description: Device to be validated
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Device'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: validate device
      tags:
      - devices
  /devices/{id}:
    delete:
      description: delete device
      parameters:
      - description: Device Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: delete device
      tags:
      - devices
    get:
      description: get device
      parameters:
      - description: Device Id
        in: path
        name: id
        required: true
        type: string
      - description: interprets the id as local_id if as=='local_id'
        in: query
        name: as
        type: string
      - description: default requesting user; used in combination with local_id (as=='local_id')
          to identify the device
        in: query
        name: owner_id
        type: string
      - description: default 'r'; used to check permissions on request; valid values
          are 'r', 'w', 'x', 'a' for read, write, execute, administrate
        in: query
        name: p
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get device
      tags:
      - devices
    put:
      description: set device; admins may create new devices but only without using
        the UpdateOnlySameOriginAttributesKey query parameter
      parameters:
      - description: Device Id
        in: path
        name: id
        required: true
        type: string
      - description: comma separated list; ensure that no attribute from another origin
          is overwritten
        in: query
        name: update-only-same-origin-attributes
        type: string
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set device
      tags:
      - devices
  /devices/{id}/attributes:
    put:
      description: set device attributes
      parameters:
      - description: Device Id
        in: path
        name: id
        required: true
        type: string
      - description: comma separated list; ensure that no attribute from another origin
          is overwritten
        in: query
        name: update-only-same-origin-attributes
        type: string
      - description: attributes
        in: body
        name: message
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Attribute'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set device attributes
      tags:
      - devices
  /devices/{id}/connection-state:
    put:
      description: set device connection-state
      parameters:
      - description: connected true/false
        in: body
        name: message
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set device connection-state
      tags:
      - devices
  /devices/{id}/display_name:
    put:
      description: set device display name
      parameters:
      - description: Device Id
        in: path
        name: id
        required: true
        type: string
      - description: display name
        in: body
        name: message
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set device display name
      tags:
      - devices
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"